Refactor(seokho/emotion): processCtrl, getApiInfo 유틸 폴더로 이동 후 emotion.ctrl에 적용 #427
