+++ app/src/models/utils/getApiInfo.js
@@ -1,0 +1,11 @@
+'use strict';
+
+function getApiInfo(method, response, req) {
+  return {
+    method,
+    response,
+    path: req.originalUrl,
+  };
+}
+
+module.exports = getApiInfo;
